Refactor in which the SortedSequence is made an actual and clean wrapper of SortedKeyValueSequence.kt. This made the InternalExtensions obsolete.
